+++ packages/app/src/renderer/src/enums/recorder.ts
@@ -308,6 +308,10 @@ export const recorderTypeOptions = [
     value: "mesio",
     label: "mesio优先",
   },
+  {
+    value: "bililive",
+    label: "录播姬引擎优先",
+  },
 ];
 
 // 录制调试等级
@@ -351,8 +355,8 @@ export const douyinStreamFormatOptions = [
 ];
 
 const qualityRetry = {
-  text: "画质匹配重试次数",
-  tip: "根据次数强制查询匹配画质，在未选择原画的情况下，可能会导致开头漏录。匹配次数结束后如果无法匹配对应画质时会自动选择其他画质，-1为强制匹配画质",
+  text: "流匹配重试次数",
+  tip: "根据次数强制查询匹配画质及其他强制参数，在未选择原画的情况下，可能会导致开头漏录。匹配次数结束后如果无法匹配对应画质时会自动选择其他画质，-1为强制匹配",
 };
 const quality = {
   text: "画质",
@@ -363,11 +367,11 @@ export const textInfo = {
   common: {
     format: {
       text: "视频格式",
-      tip: "ffmpeg录制器：选择自动时，分段为ts，不分段为mp4，<b>我不想加flv的，总有人问，出问题别找我</b><br/>mesio录制器：不支持指定",
+      tip: "ffmpeg录制器：选择自动时，分段为ts，不分段为mp4<br/>mesio录制器：不支持指定",
     },
     recorderType: {
       text: "录制器",
-      tip: "影响最底层的录制，自动选择默认使用ffmpeg，如果使用ffmpeg经常出现问题，如时间戳跳变，卡顿，音画不同步等，可以尝试切换mesio，<b>mesio目前仅推荐小规模测试使用</b>",
+      tip: "影响最底层的录制，自动选择默认使用ffmpeg，如果使用ffmpeg经常出现问题，如时间戳跳变，卡顿，音画不同步等，可以尝试切换mesio或录播姬引擎，录播姬引擎不支持只录制音频<b>目前仅推荐小规模测试使用</b>",
     },
   },
   bili: {
@@ -402,7 +406,7 @@ export const textInfo = {
     qualityRetry: qualityRetry,
     quality: {
       text: "画质",
-      tip: "如果无法找到对应画质，会结合其他选项后选择更清晰的画质，<b>真原画画质可能导致PK花瓶</b>",
+      tip: "如果无法找到对应画质，会结合其他选项后选择更清晰的画质，<b>真原画画质可能导致PK花屏</b>",
     },
     formatName: {
       text: "流格式",