+++ packages/liveManager/src/recorder/IRecorder.ts
@@ -1,6 +1,8 @@
 import { EventEmitter } from "node:events";
 
 import type { VideoFormat } from "../index.js";
+import type { FormatName } from "./index.js";
+import type { XmlStreamController } from "../xml_stream_controller.js";
 
 /**
  * 录制器构造函数选项的基础接口
@@ -11,11 +13,12 @@ export interface BaseRecorderOptions {
   segment: number;
   inputOptions?: string[];
   disableDanma?: boolean;
-  formatName?: "flv" | "ts" | "fmp4";
+  formatName: FormatName;
   debugLevel?: "none" | "basic" | "verbose";
   headers?: {
     [key: string]: string | undefined;
   };
+  videoFormat?: VideoFormat;
 }
 
 /**
@@ -23,10 +26,10 @@ export interface BaseRecorderOptions {
  */
 export interface IRecorder extends EventEmitter {
   // 基础属性
+  type: "ffmpeg" | "mesio" | "bililive";
   readonly hasSegment: boolean;
   readonly segment: number;
   readonly inputOptions: string[];
-  readonly isHls: boolean;
   readonly disableDanma: boolean;
   readonly url: string;
   readonly headers: { [key: string]: string | undefined } | undefined;
@@ -36,20 +39,23 @@ export interface IRecorder extends EventEmitter {
   run(): void;
   stop(): Promise<void>;
   getArguments(): string[];
-  getExtraDataController(): any;
+  getExtraDataController(): XmlStreamController | null;
   createCommand(): any;
 
   // 事件类型定义
   on(
     event: "videoFileCreated",
-    listener: (data: { filename: string; cover?: string }) => void,
+    listener: (data: { filename: string; cover?: string; rawFilename?: string }) => void,
   ): this;
   on(event: "videoFileCompleted", listener: (data: { filename: string }) => void): this;
   on(event: "DebugLog", listener: (data: { type: string; text: string }) => void): this;
   on(event: "progress", listener: (info: any) => void): this;
   on(event: string, listener: (...args: any[]) => void): this;
 
-  emit(event: "videoFileCreated", data: { filename: string; cover?: string }): boolean;
+  emit(
+    event: "videoFileCreated",
+    data: { filename: string; cover?: string; rawFilename?: string },
+  ): boolean;
   emit(event: "videoFileCompleted", data: { filename: string }): boolean;
   emit(event: "DebugLog", data: { type: string; text: string }): boolean;
   emit(event: "progress", info: any): boolean;
@@ -61,7 +67,6 @@ export interface IRecorder extends EventEmitter {
  */
 export interface FFMPEGRecorderOptions extends BaseRecorderOptions {
   outputOptions: string[];
-  videoFormat?: VideoFormat;
 }
 
 /**
@@ -69,5 +74,11 @@ export interface FFMPEGRecorderOptions extends BaseRecorderOptions {
  */
 export interface MesioRecorderOptions extends BaseRecorderOptions {
   outputOptions?: string[];
-  isHls?: boolean;
+}
+
+/**
+ * Bililive录制器特定选项
+ */
+export interface BililiveRecorderOptions extends BaseRecorderOptions {
+  outputOptions?: string[];
 }