+++ packages/liveManager/src/recorder/mesioRecorder.ts
@@ -105,14 +105,14 @@ export const createMesioBuilder = (): MesioCommand => {
   return new MesioCommand();
 };
 
-export class mesioRecorder extends EventEmitter implements IRecorder {
+export class MesioRecorder extends EventEmitter implements IRecorder {
+  public type = "mesio" as const;
   private command: MesioCommand;
   private streamManager: StreamManager;
   readonly hasSegment: boolean;
   readonly getSavePath: (data: { startTime: number; title?: string }) => string;
   readonly segment: number;
   readonly inputOptions: string[] = [];
-  readonly isHls: boolean;
   readonly disableDanma: boolean = false;
   readonly url: string;
   readonly debugLevel: "none" | "basic" | "verbose" = "none";
@@ -136,14 +136,12 @@ export class mesioRecorder extends EventEmitter implements IRecorder {
     if (opts.url.includes(".m3u8")) {
       videoFormat = "ts";
     }
-    if (opts.formatName) {
-      if (opts.formatName === "fmp4") {
-        videoFormat = "m4s";
-      } else if (opts.formatName === "ts") {
-        videoFormat = "ts";
-      } else if (opts.formatName === "flv") {
-        videoFormat = "flv";
-      }
+    if (opts.formatName === "fmp4") {
+      videoFormat = "m4s";
+    } else if (opts.formatName === "ts") {
+      videoFormat = "ts";
+    } else if (opts.formatName === "flv") {
+      videoFormat = "flv";
     }
 
     this.streamManager = new StreamManager(
@@ -158,20 +156,15 @@ export class mesioRecorder extends EventEmitter implements IRecorder {
     );
     this.hasSegment = hasSegment;
     this.getSavePath = opts.getSavePath;
-    this.inputOptions = opts.inputOptions ?? [];
+    this.inputOptions = [];
     this.url = opts.url;
     this.segment = opts.segment;
     this.headers = opts.headers;
-    if (opts.isHls === undefined) {
-      this.isHls = this.url.includes("m3u8");
-    } else {
-      this.isHls = opts.isHls;
-    }
 
     this.command = this.createCommand();
 
-    this.streamManager.on("videoFileCreated", ({ filename, cover }) => {
-      this.emit("videoFileCreated", { filename, cover });
+    this.streamManager.on("videoFileCreated", ({ filename, cover, rawFilename }) => {
+      this.emit("videoFileCreated", { filename, cover, rawFilename });
     });
     this.streamManager.on("videoFileCompleted", ({ filename }) => {
       this.emit("videoFileCompleted", { filename });
@@ -187,6 +180,7 @@ export class mesioRecorder extends EventEmitter implements IRecorder {
       "--fix",
       "-H",
       "User-Agent: Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/73.0.3683.86 Safari/537.36",
+      "--no-proxy",
     ];
     if (this.debugLevel === "verbose") {
       inputOptions.push("-v");
@@ -209,8 +203,8 @@ export class mesioRecorder extends EventEmitter implements IRecorder {
       .on("error", this.onEnd)
       .on("end", () => this.onEnd("finished"))
       .on("stderr", async (stderrLine) => {
-        await this.streamManager.handleVideoStarted(stderrLine);
         this.emit("DebugLog", { type: "ffmpeg", text: stderrLine });
+        await this.streamManager.handleVideoStarted(stderrLine);
       });
 
     return command;