+++ packages/liveManager/src/utils.ts
@@ -160,7 +160,11 @@ export function isFfmpegStartSegment(line: string) {
 }
 
 export function isMesioStartSegment(line: string) {
-  return line.includes("Opening ") && line.includes("Opening segment");
+  return line.includes("Opening segment");
+}
+
+export function isBililiveStartSegment(line: string) {
+  return line.includes("创建录制文件");
 }
 
 export function isFfmpegStart(line: string) {